Fixed workbench only code

+++ scripts/Game/Systems/CRF_PlayableCharacter.c
@@ -70,12 +70,12 @@ class CRF_PlayableCharacter : ScriptComponent
 			return;
 		
 		#ifdef WORKBENCH
-		if(owner.GetPrefabData().GetPrefabName() == "{59886ECB7BBAF5BC}Prefabs/Characters/CRF_InitialEntity.et" && !EntityUtils.IsPlayer(owner) && RplSession.Mode() == RplMode.Dedicated)
+		if(owner.GetPrefabData().GetPrefabName() == "{59886ECB7BBAF5BC}Prefabs/Characters/CRF_InitialEntity.et" && !EntityUtils.IsPlayer(owner))
 		{
 			SCR_EntityHelper.DeleteEntityAndChildren(owner);
 		}
 		#else
-		if(owner.GetPrefabData().GetPrefabName() == "{59886ECB7BBAF5BC}Prefabs/Characters/CRF_InitialEntity.et" && !EntityUtils.IsPlayer(owner))
+		if(owner.GetPrefabData().GetPrefabName() == "{59886ECB7BBAF5BC}Prefabs/Characters/CRF_InitialEntity.et" && !EntityUtils.IsPlayer(owner) && RplSession.Mode() == RplMode.Dedicated)
 		{
 			SCR_EntityHelper.DeleteEntityAndChildren(owner);
 		}
